pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "speech-kws"
version = "0.1.0"
description = "Speech Commands keyword spotting experiments with multiple unknown-handling strategies."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "matplotlib>=3.7",
  "numpy>=1.24",
  "pandas>=2.0",
  "PyYAML>=6.0",
  "scikit-learn>=1.3",
  "torch>=2.1",
  "torchaudio>=2.1",
  "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
speech-kws = "speech_kws.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
